FLINK-18013: Refactor Hadoop utils to a single module
