+++ flink-filesystems/flink-hadoop-fs/pom.xml
@@ -106,6 +106,11 @@ under the License.
 				</exclusion>
 			</exclusions>
 		</dependency>
+		<dependency>
+			<groupId>org.apache.flink</groupId>
+			<artifactId>flink-hadoop-utils</artifactId>
+			<version>${project.version}</version>
+		</dependency>
 
 	</dependencies>
