+++ flink-formats/flink-sequence-file/pom.xml
@@ -43,6 +43,12 @@ under the License.
 			<scope>provided</scope>
 		</dependency>
 
+		<dependency>
+			<groupId>org.apache.flink</groupId>
+			<artifactId>flink-hadoop-utils</artifactId>
+			<version>${project.version}</version>
+		</dependency>
+
 		<!-- Hadoop is needed for SequenceFile -->
 		<dependency>
 			<groupId>org.apache.hadoop</groupId>
