+++ flink-runtime/pom.xml
@@ -65,6 +65,12 @@ under the License.
 			<version>${project.version}</version>
 		</dependency>
 
+		<dependency>
+			<groupId>org.apache.flink</groupId>
+			<artifactId>flink-hadoop-utils</artifactId>
+			<version>${project.version}</version>
+		</dependency>
+
 		<!-- optional dependency on Hadoop, so that Hadoop classes are not always pulled in -->
 
 		<dependency>
