+++ flink-runtime/src/main/java/org/apache/flink/runtime/security/modules/HadoopModuleFactory.java
@@ -18,8 +18,8 @@
 
 package org.apache.flink.runtime.security.modules;
 
+import org.apache.flink.hadoop.utils.HadoopUtils;
 import org.apache.flink.runtime.security.SecurityConfiguration;
-import org.apache.flink.runtime.util.HadoopUtils;
 
 import org.apache.hadoop.conf.Configuration;
 import org.slf4j.Logger;
