+++ pom.xml
@@ -88,6 +88,7 @@ under the License.
 		<module>flink-walkthroughs</module>
 		<module>flink-kubernetes</module>
 		<module>flink-external-resources</module>
+		<module>flink-hadoop-utils</module>
 	</modules>
 
 	<properties>
